project_poc/repo/requirements.txt
pandas
requests
pyyaml
python-dotenv
google-cloud-storage
google-auth